refactor(web): extract hooks to fix component→lib/api architecture violations
Components must not import lib/api directly. Extract useTagSuggestions and
useTokenRefresh hooks so TagInput and AuthGuard go through the hooks layer.
